chore: entity와 vo에 대한 각각의 디렉토링 생성 및 이동
