+++ app/src/main/java/com/codesoom/assignment/application/interfaces/ToyCreateService.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.application.interfaces;
 
-import com.codesoom.assignment.domain.Toy;
+import com.codesoom.assignment.domain.entities.Toy;
 
 /**
  * 상품 타입 생성에 대한 비지니스 로직을 처리한다
