+++ app/src/main/java/com/codesoom/assignment/domain/ToyProducerRepository.java
@@ -1,5 +1,6 @@
 package com.codesoom.assignment.domain;
 
+import com.codesoom.assignment.domain.entities.ToyProducer;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
