+++ app/src/main/java/com/codesoom/assignment/domain/ToyRepository.java
@@ -1,5 +1,6 @@
 package com.codesoom.assignment.domain;
 
+import com.codesoom.assignment.domain.entities.Toy;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
