+++ app/src/main/java/com/codesoom/assignment/domain/entities/ToyProducer.java
@@ -1,24 +1,19 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.domain.entities;
+
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
-/**
- * '생산자' Root Entity
- * <p>
- * All Known Extending Classes:
- * ToyProducer
- * </p>
- */
-
 @Entity
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public abstract class Producer {
+public class ToyProducer {
     @Id
     @GeneratedValue
     private Long id;
