+++ app/src/main/java/com/codesoom/assignment/domain/vos/ImageDemo.java
@@ -1,16 +1,15 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.domain.vos;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Embeddable;
-import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Embeddable
-public class Won extends Money {
-    private BigDecimal value;
+public class ImageDemo extends MediaDemo{
+    private String url;
 }
