+++ app/src/main/java/com/codesoom/assignment/domain/vos/Won.java
@@ -1,15 +1,16 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.domain.vos;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Embeddable;
+import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Embeddable
-public class ImageDemo extends MediaDemo{
-    private String url;
+public class Won extends Money {
+    private BigDecimal value;
 }
