+++ app/src/test/java/com/codesoom/assignment/controllers/ToyCrudControllerTest.java
@@ -3,10 +3,10 @@ package com.codesoom.assignment.controllers;
 import com.codesoom.assignment.ProductNotFoundException;
 import com.codesoom.assignment.application.ToyCrudService;
 import com.codesoom.assignment.controllers.dtos.ToyRequestData;
-import com.codesoom.assignment.domain.ImageDemo;
-import com.codesoom.assignment.domain.Toy;
-import com.codesoom.assignment.domain.ToyProducer;
-import com.codesoom.assignment.domain.Won;
+import com.codesoom.assignment.domain.vos.ImageDemo;
+import com.codesoom.assignment.domain.entities.Toy;
+import com.codesoom.assignment.domain.entities.ToyProducer;
+import com.codesoom.assignment.domain.vos.Won;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.*;
